+++ src/Mapbender/CoreBundle/Resources/public/mapbender.element.ruler.js
@@ -175,7 +175,7 @@
         },
         showPopup: function () {
             if (!this.popup || !this.popup.$element) {
-                this.popup = new Mapbender.Popup2({
+                this.popup = new Mapbender.Popup({
                     title: this.element.attr('data-title'),
                     modal: false,
                     draggable: true,