+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Features:
 * [ShareUrl] Fix selected layers not encoded correctly ([#1697](https://github.com/mapbender/mapbender/issues/1697), [PR#1698](https://github.com/mapbender/mapbender/pull/1698)) 
 
 
+Other:
+* Removed Popup2 ([#PR1713](https://github.com/mapbender/mapbender/pull/1713))
 
 ## v4.1.0
 Features: